refact(Tests): renamed "surefire" (maven legacy) to tests for better understanding
