+++ commons/src/test/java/fr/insideapp/sonarqube/apple/commons/tests/JUnitReportParserTest.java
@@ -1,24 +1,17 @@
-package fr.insideapp.sonarqube.apple.commons.surefire;
+package fr.insideapp.sonarqube.apple.commons.tests;
 
-import fr.insideapp.sonaqube.apple.commons.TestFileFinder;
 import fr.insideapp.sonaqube.apple.commons.TestFileFinders;
-import fr.insideapp.sonaqube.apple.commons.surefire.SurefireReportParser;
+import fr.insideapp.sonaqube.apple.commons.tests.JUnitReportParser;
 import org.junit.Test;
-import org.sonar.api.batch.fs.FileSystem;
-import org.sonar.api.batch.fs.InputFile;
-import org.sonar.api.batch.fs.internal.DefaultInputFile;
 import org.sonar.api.batch.fs.internal.TestInputFileBuilder;
-import org.sonar.api.batch.sensor.SensorContext;
 import org.sonar.api.batch.sensor.internal.SensorContextTester;
-import org.sonar.api.batch.sensor.measure.Measure;
 import org.sonar.api.measures.CoreMetrics;
-import org.sonar.api.measures.Metric;
 
 import java.io.File;
 
 import static org.assertj.core.api.Java6Assertions.assertThat;
 
-public class SurefireReportParserTest {
+public class JUnitReportParserTest {
 
     @Test
     public void collect() {
@@ -28,7 +21,7 @@ public class SurefireReportParserTest {
                 new TestInputFileBuilder("", "SQAppTests/SQAppTests.swift").setLanguage("swift").build());
 
         SensorContextTester context = SensorContextTester.create(new File("src/test/resources"));
-        SurefireReportParser parser = new SurefireReportParser(context);
+        JUnitReportParser parser = new JUnitReportParser(context);
         parser.collect(new File("src/test/resources/tests"));
 
         assertThat(context.measure(":SQAppTests/SQAppTests.swift", CoreMetrics.TESTS).value()).isEqualTo(4);
